requirements.txt
google-cloud-vision==3.4.4
google-generativeai==0.3.2
Pillow==10.3.0
python-dotenv==1.0.0